+++ apps/ospTestSuite/test_geometry.cpp
@@ -1,4 +1,4 @@
-// Copyright 2017-2020 Intel Corporation
+// Copyright 2017-2021 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
 #include "test_geometry.h"
@@ -85,7 +85,7 @@ void SpherePrecision::SetUp()
   distant.setParam("angularDiameter", 1.0f);
   AddLight(distant);
 
-  cpp::Light ambient = ospNewLight("ambient");
+  cpp::Light ambient("ambient");
   ambient.setParam("intensity", 0.1f);
   AddLight(ambient);
 }